cluster_analysis: fix error with similar articles
check if atleast one index in indices_with_similarity_greater is not present in
indices_of_same_source.

comparing equality doesnot work because,
ex. indices_with_sim_greater = [2] and indices_of_same_source = [2,3]
in this case even though they are not equal but still there is not index
with similarity greater with other source.

Hence we need to check if indices_with_sim_greater is not equal or is
not a subset of indices_of_same_source